fixxing lesson_2

+++ Lesson_2/GeneralizedHarmonic.java
@@ -1,8 +1,8 @@
 public class GeneralizedHarmonic {
 
     public static void main(String[] args) {
-        double i = Double.parseDouble(args[0]);
-        double r = Double.parseDouble(args[1]);
+        double i = Integer.parseInt(args[0]);
+        double r = Integer.parseInt(args[1]);
         double t = 0;
         for (double n = i; n > 0; n--) {
             t += (1 / Math.pow(n, r));
